Implement External Decision Agent (EDA) builtin policy
Implements a very simple (builtin) policy that simply looks at cpu.*
container annotations (ofthe CRI container object) in the
cri-resource-manager namespace (i.e.
cri-resource-manager.intel.com/cpu.*). If any annotations are found, the
cpusets within are combined and and the cpuset of the container is set
accordingly.

For example, having:
 cri-resource-manager.intel.com/cpu.gold-1: "2"
 cri-resource-manager.intel.com/cpu.gold-2: "4"
will limit the cpuset workload to CPUs 2 and 4.