Implement RDT support
This patch implements RDT support. The resource manager now has a common
confugration of a set of RDT classes (or resctrl CTRL groups) which the
policies are able to assign containers to. This RDT configuration is
dynamically updatable via the resource manager node agent
(cri-resmgr-agent). Resource manager has a built-in default
configuration containing three classes corresponding to the Pod QOS
classes of Kubernetes. However, this default configuration is sort of
sample configuration that specifies 100% L3 allocation for all classes,
making all classes equal in terms of RDT. The set of classes can be
freely specified in the configuration, but, one must ensure that classes
required by the active policies have been defined, and, that the
maxmimum number of classes (CLOSes) supported by the hardware is not
exceeded.

This patch integrates RDT support to the static policy. The policy
simply assigns each container to an RDT class corresponding its Pod QOS
class (i.e.  'Burstable', 'BestEffort' or 'Guaranteed').

The patch implements a soft autodetection for RDT so that the RDT
support in cri-resmgr is simply disabled if cri-resmgr is not able to
detect the resctrl filesystem (i.e. the user space interface for RDT).
However, cri-resmgr errors out if the user explicitly specifies a mount
path via the '-resctrl-path' command line flag but the given path does
not contain a valid, accessible resctrl file system. RDT support can be
disabled by using the '-no-rdt' command line option.