pkg/version, cri-resmgr: add linker-aided autoversioning.
Add basic support for automatic git->binary version propagation.
The necessary Makefile glue is not in place yet, coming in the
next patch. Once in place, version (annotated git tag) and build
(git SHA1) information will be automatically propagating to the
binaries built and can be queried with the --version command line
option.