cache: add container affinity and anti-affinity support.
Using the new pod level affinity and anti-affinity annotations
one can assign signed weights to pairs of containers running on
the node.

This should allow policies to let users express soft 'pull' and
'push' preferences between containers. Containers that pull each
other should be put close together and containers that push each
other should be put apart from each other wrt. HW-topology.

Of course, policies are free to define/use any other semantics
for affinity as they see fit.

Here is an example of an annotation which defines an affinity
from container4 to container1 and 3, and anti-affinity to
container2.

metadata:
  name: "test-affinity"
  annotations:
    cri-resource-manager.intel.com/affinity: |
      container4: [ container1, container3 ]
    cri-resource-manager.intel.com/anti-affinity: |
      container4: [ container2 ]

In reality affinities are more expressive than this and hence
a bit more complex.

A single affinity for a container consists of an optional scope
expression, a match expression and an int32 weight. The scope
defines which containers are in the scope (IOW evaluated) for the
affinity. If the scope is omitted, the affinity scope is the pod
of the container. The match expression defines which containers
within the scope the affinity applies to, and the weight specifies
the weight assigned to the pair of containers.

An expression consists of a key, an operator and a set of values.
It represents the logical operation <key op values>. The possible
operations are currently Equals, NotEqual, In, NotIn, Exists and
NotExist. These implement tests for equality, inequality, set mem-
bership, negated membership, key existence and non-existence.

Here is what the above example really is once the omitted defaults
are filled in after parsing the annotation.

metadata:
  name: "test-affinity"
  annotations:
    cri-resource-manager.intel.com/affinity: |
      container4:
      - match:
          key: io.kubernetes.container.name
          operator: In
          values:
          - container1
          - container3
        weight: 1
      - match:
          key: io.kubernetes.container.name
          operator: In
          values:
          - container2
        weight: -1