pkg/config: support for more generic configurability.
Add a configuration package which provides a unified, more
generic mechanism for receiving configuration data from a
number of sources. Like our previous configuration handling
code this is also built on top of the standard flag package.
Therefore the switch to using it should be straightforward,
at least in principle.

The configuration abstraction splits runtime configuration
data into a set of modules. Each module has its own set of
variables, namespaced by the name of the module. A module
could be any self-contained logical unit of functinality,
typically a package. Modules declare variables (configuration
data) in an identical manner to the flag package.

In addition to the basic (flag package) functionality the
config package supports

- taking configuration from multiple sources, currently
  o command line
  o YAML or JSON file
  o an(y) external source
- multiple configurations
- configuration change check/notifications
  o both per configuration and per module
  o with support for
    - rejecting configuration changes
    - automatic rollback of rejected configuration