agent: move node configuration selection to the agent side.
Move the logic for picking configuration data for the node
to the agent. 3 ConfigMaps are tracked for picking node
configuration, in decreasing order of preference:

  1. node-specific: cri-resmgr-config.node.<node-name>
  2. group-specific: cri-resmgr-config.group.<group-name>
  3. default: cri-resmgr-config.default

If none of these yields any data the node configuration is
the empty configuration.

A node is assigned to a configuration group by setting the
dedicate group-label on the node. This label is by default
cri-resource-manager.intel.com/group and can be changed by
the --label-name command line option.