agent, resource-manager: add error field to config update reply message.
Return configuration update errors as a field in the reply message.
This allows the agent to differentiate between errors in delivery
vs. errors in applying an update.

For delivery errors the agent will keep retrying indefinitely, sending
an update request at periodic intervals. For errors in applying an
update the agent will give up immediately and only try sending a new
update when the configration is updated the next time.