policy: pass cache early, remove PostStart, nuke Implementation.
We'll change configuration handling to get the initial
configuration applied before we fully start up. Pass the
cache as early to policies as possible, already in the
constructor.

Remove policy-level PostStart(). Equivalent actions are
now supposed to be performed by resource controllers.

Get rid of policy.Implementation for registration. It was
only used to pass metadata about policies up to the gene-
ric policy layer but was mostly confusing people.