policy: pass cache early, nuke PostStart, SetConfig, Implementation.
We'll change configuration handling to get the initial
configuration applied before we fully start up. After
that policies could boot right to the latest up-to-date
configuration. But only if they had access to the cache
early enough, so pass the cache as early to policies as
possible, already in the constructor.

Remove obsolete policy-level PostStart() which was only
used to trigger RDT control directly bolted into some
of the policies. Also get rid of obsoleted policy-level
SetConfig.

Get rid of policy.Implementation for registration. It was
only used to pass metadata about policies up to the gene-
ric policy layer but was mostly confusing people.