rdt: simplify l3 configuration
This patch tries to simplify the configuration of the RDT L3 allocation
and aims to make it slightly more comprehensible and user friendly.
Instead of having independent configuration sections (l3Schema,
l3CodeSchema, l3DataSchema), we now have separate 'unified', 'code' and
'data' settings under the 'l3Schema' key. E.g.
    resctrlGroups:
      Guaranteed:
        l3Schema:
          all:
            unified: 90%
            code: 100%
            data: 80%
Moreover, 'unified' must always be specified, and, 'code' and 'data' go
together so that none or both of them must be specified (only specifying
one of them creates an error). These requirements were added in order to
reduce the risk of accidental misconfiguration.

Also. the fallback/failure logic has been simplified. Now the
configuration only fails if 'l3Schema' is specified but L3 CAT is not
available and the 'rdt.config.options.l3.optional' is set to false. In
addition the 'l3' setting (which is now mandatory) is used as a
default/fallback value for 'code' and 'data' settings.