policy: add ReallocateResources() to experiment with reblancing.
Add ReallocateResources to the generic policy layer.
It is primarily intended to be used for rebalancing
container allocations.

ReallocateResources first releases the resources of
all given containers using ReleaseResources. Then it
allocates them again by calling AllocateResources for
each in the order the containers are found in the
given slice.