policy: add Rebalance(), stub/dumb backend implementations.
Add policy.Rebalance() and corresponding stub or dumb
implmentations to the existing policy backends.

Rebalance() is the primary mechanism for rebalancing or
optimizing the resource allocations for the current set
of running containers. The resource manager will trigger
rebalancing periodically. It is up to the policy backend
implementation what kind of rebalancing strategy it uses.

The current API presents no way for the upper layer to
restrict which containers are in the scope of rebalancing.
Regardless, implementations should honor the implied QoS-
semantics of pods/containers: containers in Guaranteed QoS
pods are not to be disrupted/touched.

This patch adds stubs to all but the topology-aware policy
backend. The topology-aware policy Rebalance() does a dumb
reallocation of all non-Guaranteed QoS containers in quasi-
random order (as returned by iterating the map in cache).