rdt: allow configuration of the root resctrl group
Now the reserved class name 'SYSTEM_DEFAULT' can be used in our rdt
configuration to specify the allocation for the "root" resctrl group of
the system (located in the root of the resctrl mount point). CRI-RM does
not do any modifications to the root resctrl group if SYSTEM_DEFAULT
class is not specified in the config. Thus, the existing default
functionality does not change, but, this gives the user the possibility
to configure the root group using CRI-RM.