cache: add AlwaysTrue (affinity expression) operator.
Add a new AlwaysTrue operator which always evaluates to
true. It can be used to denote a 'global' scope of containers,
a scope containing all containers running on the node.

For instance, the affinity

  Affinity{
    Scope: &Expression{
      Op: AlwaysTrue,
    },
    Match: &Expression{
      Key:    "xyzzy",
      Op:     In,
      Values: [ "foo", "bar" ],
    },
    Weight: 5,
  }

will cause an affinity with weight 5 towards all containers
(AlwaysTrue scope) which a label 'xyzzy' set to either 'foo',
or 'bar'.