cache: add support for defining implicit affinities.
Add support for registering implicit affinities. An implicit
affinity is defined by a container eligibility test function
and an affinity expression. When affinities for a container
are looked up by the cache, all those implicit affinities get
included for which the eligibility function called with the
Container returns true.

Implicit affinities can be used to inject policy-specific
affinities to all or a selected subset of containers.