cache: make explicit anti-affinity implicitly negative.
If a weight is given for an anti-affinity annotation, always
multiply it by -1. This seems more intuitive and it makes anti-
affinity and affinity annotations identical except for the
annotation key name. It also results in a more logical/symmatric
behavior with a negative affinity becoming an anti-affinity and
a negative anti-affinity becoming an affinity.
it results in a more logical behavior
of