resource-manager: add event-/metrics-processing, rebalancing.
Add event processing and periodic rebalancing. Event processing
pulls events from the metrics data collector/preprocessor. It
should produce data which the policy-level rebalancing code can
use to

  - decide if rebalancing is warranted at all
  - choose a compomise between improvement vs. disruption

The periodic rebalancing simply triggers the rebalancing algorithm
of the active policy. It is up to the policy to decide if rebalancing
is needed and to which extent it is performed if at all. The policy
should find a reasonable compromise between a fully optimal re-
allocation and disruption to the running containers caused by the
rebalancing actions.

This patch only adds the necessary bare minimum plumbing (apart from
the PoC AVX-tagging for testing). There is no code for digesting and
providing data for containers yet.