pkg/sysfs: kludge to exclude test data from (zip archives in) go get.
The zip archive code used in the go toolchain can't handle files
with a colon (:) in their name, even if those files are not used
by the toolchain in any way.

As a workaround, we force such files to be excluded from the parent
module by dropping an empty go.mod file above their location in the
source tree.

Workaround spotted in https://github.com/golang/go/issues/26672.