sysfs: split out topology to a subdir (and subpackage) of its own.
We have a sibling project that would like to pull in some of the
topology hint calculation functionality from pkg/sysfs but would
like to do so with minimal extra collateral. By splitting it out
to a subpackage of its own it is now practically self-contained,
importable without extra cruft.