log: add log.Delay() helper for delayed evaluation.
Add a helper function for delayed evaluation of expensive
arguments to Logger.Debug(). You can use it to avoid the
cost of evaluating really expensive arguments (complex
formatting) to Logger.Debug when debugging is not enabled.

You can use it like this:
  var data SomeType
  ...
  log.Debug("%T: debug-logged with heavy lifting: %s",
      data, log.Delay(func() string { return heavy_lifting(data) }))
  ...

With the above construct, heavy_lifting() will only be called if
debugging is enabled for the logger instance 'log' above.