pkg/cgroupstats: don't use DescribeByCollect for describing metrics.
prometheus.DescribeByCollect() can only be used by collectors
that always produce data for all the metrics they ever collect.
This collector is not like that. It does not produce any data
if there are no containers (and related cgroups), which can be
the case when we start up. Hence change the collector to explic-
itly describe all of its metrics without having to collect any.