Rudimentary support for CPU prioritization
This patch adds support for per core priority in the cpu allocator used
by static, static-plus and topology-aware built-in policies. At this
point there is only support for two stages of priority (low and high).
The new allocator heuristics added in this patch are straightforward:
depending on the caller's guidance, the allocator always simply prefers
or tries to avoid high-priority CPUs.

Detection of high priority CPUs is currently based on the base_frequency
attribute. If CPUs with different base frequencies are detected, all
higher frequency CPUs (higher than the lowest detected base frequency)
are flagged as high priority.