rdt: add rdt.Control-less version of GetClasses().
We don't have now easy access to rdt.Control (hidden
in the resource manager RDT controller). However, we
do need the ability to list RDT classes for external
introspection. Hence, provide a version of GetClasses()
which does using the last loaded configuration instead
of the one stored in rdt.control.