resource-manager: start policy, set bypass for null policy.
Initialize/start the policy layer normally even if the active
policy is the null policy. Set a bypass check function for
the server to bypass local policy processing in this case
altogether. Additionally, omit initial synchronization with
the runtime and starting the event/metrics processing loop
if we bypass local processing.