rdt: un-export Control interface
The resctrl API is essentially a singleton on the system level, thus,
there should be no reason to have multiple RDT separate control
interfaces in CRI-RM, either. This drops the 'Control' interface and
replaces that with "singleton" functions.

Also, add minimal unit test for the public interface, testing that an
uninitialized rdt interface works as expected.