Makefile: fix discovery of protoc command
On systems where 'command' is only available as a shell built-in. Adding
the semicolon causes gmake to invoke a true shell environment and the
command succeeds. Without this fix protoc utility is not found correctly
and every invocation of 'make' causes annoying
"make: command: Command not found"
to be printed in the terminal.