memtier: implemented cold start support.
The idea with cold start is that a container can first be assigned PMEM
memroy and the DRAM controller to be added to the container memset only
later. This means that the initial memory allocations will be done from
PMEM, causing there to be less need for memory demotion later in the
runtime. This can be beneficial if the working set size is small.