policy: make Bypassed() a member of the policy interface
Fixes a segfault caused by config updates where the configuration of the
policy layer and the policy instance get out-of-sync. This happens
because we do not currently handle run-time configuration updates of
the upper policy layer. If we start with an empty config, (i.e. null
policy) no policy backend gets initialized. But, after a config update
the active policy in the configuration might have changed to something
else. Now, Bypassed() returned false (because it was looking at the
config) but in reality we did not have any policy backend. This caused a
segfault at request processing as we tried to do policy decisions with a
nil backend.

Even for the future, when config updates are fixed/supported, it is
probably safest to get this information directly from the policy
instance, rather than from its configuration.