resource-manager: always setup node agent connection
Always setup connection to the node agent as it serves other functions
than just config handling. Formerly the agent connection was not
initialized if --force-config was used. This was problematic, as e.g.
static-pools policy requires services (other than config) from the agent
and refused to work when forced config was used.