cache: add support for opting out from topology hint generation.
Using the new 'topologyhints' resource manager annotation
key it is now possible to opt out from automatic topology
hint generation on a per pod or container basis.

Use this annotation to opt out a full pod:
  annotations:
    topologyhints.cri-resource-manager.intel.com/pod: "false"

Use this annotation to opt out container 'foo' in the pod:
  annotations:
    topologyhints.cri-resource-manager.intel.com/container.foo: "false"

Currently topology hint generation is enabled by default, so
using the annotation as opt in (setting it to "true") should
have no effect on the placement of containers of a pod. This
might change in the future however.