memtier: (hopefully) better affinity score evaluation.
Calculate affinity for every node as a combination of
affinities of the nodes on the path from the node to
the root and the nodes in the subtree under the node.

The combined affinity for node n is Sum_x(A_x/D_x),
where for every node x, A_x is the affinity for x and
D_x is 1<<(number of links from node to x). IOW, the
effective affinity is the sum of the affinity of n and
the affinit of each node x of the above mentioned set
diluted proprotionally with the distance of x to n.