.github/workflow: add release-tests.yaml.
Add github workflow to run release tests. This workflow
gets triggered for
  - pull requests to a branch matching 'releases/**'
  - tags matching v*
  - push to a branch matching 'releases/**' or '**/release-tests'

The workflow includes a conditional tmate action which sets
up a tmux session on the worker. This action gets triggered
for pushes both to and from any branch that contains 'debug'
in its name. If the session is created it can be remotely
accessed using ssh or a web console, the details of which
can be found in the action's logs.