docs: show version in the site sidebar
Add --site flag in scripts/build/get-buildid for printing a version
label for documentation. This version is either the majmin of a released
version (e.g. '0.4') when building a tag (e.g. 0.4.1) or simply 'devel'
when building an untagged commit. The thinking behind this is that we
would maintain multiple versions of the documentation - one for each
release branch (release-majmin) and one for the tip of the master
branch.

The version handling in git-buildid is quite bash-ish but we probably
don't care about portability here.