scripts: add script for updating gh-pages branch
The new script builds documentation from the currently checked-out
commit and update the gh-pages branch. In more detail, it updates a
subfolder (in the gh-pages branch) corresponding the major version  of
the software. The version used here is parsed from the GITHUB_REF
environment variable and defaults to 'devel' if it is not defined. Also,
for master branch use subfolder name 'devel' instead of using the branch
name (which would be the case for other non-release branches). The
"version subfolder" can also be given directly from the command line,
e.g. for testing purposes.

update-gh-pages.sh: map 'master' to 'devel'