docs: archive documentation for tagged releases
Use a separate 'releases' subdir for the documentation archive. The root
of this subdir is effectively a separate site built with sphinx. Site
configuration lies under 'docs/releases' and it is only being (re-)built
on the master branch. This is to ensure that master branch is in control
of the releases page and no older release branches (or tags) can mess it
up. The actual list of releases shown on the page is dynaically
generated, in a similar fashion to the version menu. The releases/
directory is supposed to have releases.js which lists all archived
releases. Unlike the actual html content, this "release metadata" will
be updated on every execution of update-gh-pages.sh.