docs: create file references to github repo
When referencing files in this repo, create links to our github repo
instead of making a local copy.

Also, fixes handling of file references. When running through missing
references, only ignore file references with an absolute path. These
will reference the repo root directory and will be turned later on to
references to GitHub, directly. Missing references with a relative path
raise an error (as they should because the file really is missing).

Fix the broken references in the documentation, too, so that the
documentation builds without errors.