static-pools: re-load legacy config on config update
Change the handling of pools configuration.

First, respect the legacy config path (ConfDirPath, ConfFilePath)
specified in the dynamic configuration. If these settings change we
try to (re-)load the legacy config from the given location.

Second, the pools config priority should now be more consistent and
evident. The priority in descending order is:
1. dynamic configuration
2. leagcy config file
3. legacy config directory structure