memtier: rebuild pools if the new config requires it.
If the allowed or reserved resources have changed, we need to
rebuild our pool hierarchy (with the updated constraints) and
apply the grants with the new pools. This we do in exactly the
same way as during startup, when we reload the grants from the
cache:
  1. try to reinstate the grants verbatim (no resource changes)
  2. if this fails, try to reallocate all containers
  3. if this fails too, reject the configuration with an error
If we end up rejecting the configuration, we restore the original
state of the policy which we save before rebuilding the pools.