rdt: use root class as a fallback to missing classes
Assign container to the system root class if the class assignment fails.
The most probable reason is that the class specified for it is missing.
Otherwise the unassigned container(s) will inherit the resctrl group of
the container runtime itself which we want to avoid.  The system root
class is guaranteed to exist and is the obvious fallback default.