pkg/sysfs: implement adaptation for SST
Add rudimentary read-only support for interfacing with the kernel
control interface of Intel Speed Select Technology (SST).

This functionality is not using the sysfs interface but the sysfs is the
most convenient location for the code because of the high dependency to
topology detection features.

The implementation of the ioctl interface is based on Linux kernel
headers, the source code of the intel-speed-select tool under Linux
kernel source tree [1] and SST mailbox interface description [2]. The
interface type definitions are auto-generated from Linux kernel sources.
The required minimum Linux kernel version is v5.6.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/tree/tools/power/x86/intel-speed-select?h=v5.10.11
[2] https://github.com/intel/CommsPowerManagement/blob/5155b11e2ea66c8989a8368ec671af9243568e7c/intel_sst_os_interface/mailbox.md