cpuallocator: support SST
Discover CPU priority set via Intel Speed Select Technologies (SST). SST
Turbo Frequency (SST-TF), Core Power (SST-CP) and Base Frequency
(SST-BF) are supported. The heuristics of determining CPU priority are
simple and we do not try to analyze the configuration parameters deeper.
That could be improved in the future even though this path is covered
with potholes and pitfalls because of large configuratio space, overlap
with HWP (khobs in cpufreq/intel_pstate driver) and platform
differences.

The current implementation only uses one source of information (per CPU
package) for determining CPU priority: first SST if that has been
enabled, falling back to cpufreq if SST is not supported or is disabled.