e2e: update dynamic page demotion tests
This test was written with the assumption that every pod runs on its
own socket in the 4-socket system, and is pinned to the nearest DRAM
and PMEM memory. Now that socket #0 is reserved for kube-system and
will not run test workloads, one of the four workloads will be placed
on the root node. This increases the scope of this test, as we are
measuring demotion speed over many DRAM and PMEM nodes. This patch
updates the tests to correctly calculate total PMEM pages from many
PMEM nodes.