e2e: lib: Use proper locale for bc to work
The current time returned by bash variable EPOCHREALTIME can contain
either "." or "," separator between seconds and microseconds (depending
on used locale), and because that value is feed directly to bc, then it
is possible that bc command will not understand the value and can return
"(standard_in) 1: syntax error"
if the current locale uses "," separator.

In order to fix this issue, set LC_NUMERIC=C so that bash uses always "."
when feeding the data to bc command.