e2e: Enable cri-o in centos-8
If user has not specified cri-o sources, then system specified
cri-o package is used but it is not enabled by default in centos-8.
This means that in second boot of the VM, the kube environment does
not start properly because cri-o is not started automatically.
Fix this by enabling cri-o in systemd so that it is started at boot.