e2e: Allow proxy setup even if VM_IP is set
Normally proxies are not setup if the VM_IP variable is set.
The VM_IP is typically set if we are not using govm but instead
the use real host. For real host setups, the default is not to
setup proxies as it is assumed that proxies are already set there.
If user specifies "setup_proxies=1" when starting the run.sh script,
then the proxies are setup in host even if VM_IP is set.