resource-manager,server: better socket probing.
Reduce active socket test logic by testing if connecting
to the socket is accepted without going through the full
gRPC stack. The socket is known to be in listening state
if connections succeed. The socket is known not to be in
listening state if connections fail with ECONNREFUSED or
ENOENT. Otherwise the test fails with an error.