pkg/cri/server: implement ingress CRI conversion.
Implement v1/v1alpha2 CRI protocol conversion on the
ingress where we act as a CRI server for a client.
Do the protocol conversion by marshalling v1alpha2
then unmarshalling it into v1 for requests and the
other way around for responses.

Although this is slower than relying on the versions
to be struct-compatible and doing an unsafe cast, it
is much safer should the versions divert.