resmgr: rediscover container resources on startup.
Try rediscovering container resources as part of our state
synchronization with the runtime during startup. We do this
by sending for each container a verbose CRI ContainerStatus
request, digging out the OCI Spec from the runtime-specific
extra info in the response, then reconstructing a synthetic
CRI LinuxContainerResources instance from this. Then we use
that to estimate the container resource requirements, just
like we do when processing a CreateContainerRequest.

This is necessary for properly rediscovering containers if
the stored cache has been nuked/reset.