sysfs: check and report bogus NUMA meminfo entry.
On some HW and kernel combinations we've seen more free than total
memory being reported. This causes exorbitant usage of memory being
reported which later can cause failures in policies which trust and
rely on this information.

Report an error (and bail out immediately) when such a problem is
detected.