control: add cpu controller
For controlling cpu parameters such as cpufreq attributes. This patch
adds a stub controller that is missing the functionality for actually
configuring the cpus (cpufreq).

It implements dynamic configuration updates as well as internal state
tracking (stored in cri-rm cache) that stores the information about
which classes cpus are assigned to (if any).

The actual controller hook functionality is also currently a stub and it
does not rely on any container lifetime hooks.

An example configuration:

cpu:
  classes:
    high-prio:
        minFreq: 2000000
        maxFreq: 5000000
    low-prio:
        minFreq: 1000000
        maxFreq: 1500000