sst: Pass around pointer to SST info struct
Earlier code manipulated full SST info in CPU discovery, but this
is not needed as we can just use the pointer to SST info in the code.
This clarifies the usage, as the SstPackageInfo struct contains a slice
which can confuse the user if SstPackageInfo is not passed around as a
pointer value. This also saves some tiny amount of memory.