control/cpu: add uncore frequency controls
Support uncore frequency limits in the cpu controller. This patch adds
new 'uncoreMinFreq' and 'uncoreMaxFreq' fields to the cpu class
configuration.

Uncore frequency limits are only controllable on the cpu die level in
the HW. When resolving the per-die effective uncore frequency the
controller prefers performance i.e. selects the highest requested value.
For every die it selects the highest value (for both min and max) of all
cpu classes that are effective on that die.

An example configuration with uncore limits:

    cpu:
      classes:
        high-prio:
            minFreq: 2000000
            maxFreq: 5000000
            uncoreMinFreq: 2000000
            uncoreMaxFreq: 2400000
        low-prio:
            minFreq: 1000000
            maxFreq: 1500000
            uncoreMinFreq: 1200000
            uncoreMaxFreq: 1800000