docs: put goresctrl version in sphinx config
Make the goresctrl version that we refer to in the docs configurable in
the sphinx config. This makes it easier to maintain (keep in sync
with the version we actually depend on). A future improvement would be
to automatically get the correct version e.g. directly from go.mod.

The patch also enables myst "substitutions" extension to make this work.