e2e: Fix the kernel cmd line handling in Fedora
In Fedora, the manipulation of /etc/default/grub was not enough
as the /boot/loader/entries/<kernel>.conf file needs to be updated
too. The grubby tool is able to update all the needed files in Fedora
so use it. For other distros, use the existing distro-set-kernel-cmdline()
function like before.