e2e: Add support for Github self hosted action runner
The scripts in this directory will:

* Provision a Vagrant based VM to be used by the e2e tests
* Run a self-hosted action runner which will setup a VM for the
  tests, and then discard the VM after the tests have finished.

You need first to provision the VM that runs the e2e tests.
Then that VM is launched by Github self hosted runner script to
execute the tests. The self hosted runner is triggered to run the
tests from Github workflow actions yaml file.