test/e2e/run.sh: add qemu_cpumem='...' option.
Add $qemu_cpumem option which can be used to directly specify
qemu CPU and memory configuration options for the VM to being
created. It is mutually exclusive with $topology.

Among other things this new option can be used to bootstrap
a KVM-enabled test VM (if nested virtualization is enabled on
the host), for intstance using this commnand line:

topology='' \
qemu_cpumem="-cpu Haswell-noTSX-IBRS,vme=on,ss=on,vmx=on,\
pdcm=on,f16c=on,rdrand=on,hypervisor=on,arat=on,tsc-adjust=on,umip=on,\
md-clear=on,stibp=on,arch-capabilities=on,ssbd=on,xsaveopt=on,pdpe1gb=on,\
abm=on,ibpb=on,ibrs=on,amd-stibp=on,amd-ssbd=on,skip-l1dfl-vmentry=on,\
pschange-mc-no=on -smp cpus=8 -m 8G" \
vm=test-vm \
distro=fedora \
k8scri='containerd' \
./test/e2e/run.sh interactive