config: refactor/rework config handling.
In preparation for enabling runtime policy switching,
rework configuration handling internals. The original
string-indexed map of configuration fragments, stored
essentially as a map[string]interface{}, is now gone.
It is replaced by a tree of fragment struct pointers,
stitched together according to the registered config
hierarchy using reflect.

The requirements for a registered fragment struct are
now more formal, described by interfaces. The minimal
interface defines Reset() and Describe() which replace
the registered default value instance and description.
An additional and optional Validate() replaces the
original attachable config notification callback.

In the refactored implementation it is not possible to
attach a notificaton callback to a 'foreign' part of
the configuration tree any more. These notifiers were
used as a kludge to trigger the activation of updated
configuration, one piece at a time, without global co-
ordination, in a not too incorrect partial order. Now
activating an updated configuration is explicit and in
cases where activation does not have any data or other
dependency to anywhere outside the configured entity,
activation can be done in connection with successful
validation.