topology-aware: ignore kubelet topology 'pseudo-hints'.
Ignore any topology pseudo-hints from the 'kubelet' provider.
Such hints are present only if the kubelet has the CPUManager
enabled (with the static policy), which is not a recommended
setup. Even if we run in such an environment, the original
allocations should normally have no effect on our decisions
and should be simply ignored.