resource-manager: remove 'null' pseudo-policy.
Remove the 'null' pseudo-policy and the related bypass code
paths which leave part of the CRI relaying and intercepting
infra uninitialized. Use the no-op 'none' policy by default.

This setup should be a more reasonable default. It should
allow us to enable/handle switching from the default 'none'
policy to any other one during a configuration update. And
that should make it easier to correctly set up the initial
configuration when we expect the real one to come from the
agent, eliminating the need for a fallback config with the
final policy already enabled.