resource-manager: Check data when setting / loading configuration
The conf.Data must be non-null when receiving it from agent,
we must not accept invalid data into the cache. This could happen
because of timing when cri-rm and agent are started at the same time
(like when using DaemonSet) and agent giving a not yet ready configuration
to cri-rm which causes invalid policy settings to be used.