github: update release instructions
Instruct to use 'cross-tar' make target for building the "distroless"
binary tarball. This builds inside a container (like cross targets)
which means that we have more reproducible builds, e.g. using the golang
version specified in the Makefile instead of using whatever happens to
be installed on the host system (if any).