Makefile: add cross-tar to cross-packages
Also build the "distroless" binary tarball with the corss-packages
target similar to the deb and rpm packages.

Also, change packaging-tests target to depend on cross-packages so that
it also tests the binary tarball. Even if it currently only tests the
buildability of the tarball as there is no actual e2e test to install
and verify it. Adding a packaging e2e-test for the binary tarball is
left as a future exercise.