pkg/agent: add support for HealthCheck queries.
Allow limiting the scope of a HealthCheck query by specifying
a query string. Add a constant for querying the status of the
last configuration update pushed to the resource manager.

When receiving a ConfigStatus query, fetch the result of the
last config update pushed to the resource manager and reply
according to this. Otherwise keep responding as before. This
should allow us to run the original liveness probe unmodified.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>