agent, resmgr: nuke agent.GetConfig().
[cherry-picked #986/9c4ccf3]

We started propagating the result of the last config update
from the resource manager through the agent and the agent
probe back to the control plane. Therefore, stop querying
the agent for configuration data on the startup path. If we
failed to apply the queried configuration, there would be
no way to relay that information back to the control plane,
unless we added a dedicated interface on the agent side of
the gRPC API, which we rather don't do.

Since we stopped using agent.GetConfig(), remove it from
the protocol, agent and resmgr.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>