github: take golang version from go.mod
Fix the verify workflow to actually install the version of Go specified
in go.mod. Previously, because of the yaml parsing peculiarities go v1.2 was
installed.

Also changes the order of steps so that go.mod is available.

NOTE: This expects that the practice from now on is to keep the Go version
in go.mod and Makefile in sync.

(cherry picked from commit b5128e8c436dd6e151d53cd007da2c7be10866ab)